refactor(#183): adapter/out 구조 정리 및 mapper 패키지 이동
