+++ src/main/kotlin/team/mozu/dsm/application/service/article/QueryArticleDetailService.kt
@@ -3,7 +3,7 @@ package team.mozu.dsm.application.service.article
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import team.mozu.dsm.adapter.`in`.article.dto.response.ArticleResponse
-import team.mozu.dsm.adapter.out.article.persistence.mapper.ArticleMapper
+import team.mozu.dsm.adapter.out.article.mapper.ArticleMapper
 import team.mozu.dsm.application.exception.article.ArticleNotFoundException
 import team.mozu.dsm.application.port.`in`.article.QueryArticleDetailUseCase
 import team.mozu.dsm.application.port.out.article.QueryArticlePort
